Initial implementation of gameobjects and composite pattern + refactored camera/camera controller to make it more modular
